==> generative_math_model/01_create_math_corpus.ipynb <==
import os
import pandas as pd

project_folder = "diygenomics-projects"
sub_category = "math"
input_file = 'extracted_annotated_math.csv'
index_col = 'uuid'

data_path = os.getenv('DATA_PATH')
file_path = lambda *args: os.path.join(data_path, project_folder, sub_category, *args)
root_dir = file_path()

dataframes = []

for dirpath, dirnames, filenames in os.walk(root_dir):
    if os.path.basename(dirpath) == 'mathpix':
        csv_file = os.path.join(dirpath, 'extracted_annotated_math.csv')
        if os.path.isfile(csv_file):
            df = pd.read_csv(csv_file, index_col=index_col)
            df['folder_name'] = os.path.basename(os.path.dirname(dirpath))
            dataframes.append(df)

combined_df = pd.concat(dataframes)

combined_df['math']

combined_df.to_csv(file_path('combined_extracted_annotated_math.csv'))

! ls {file_path()}
